Unify common code in distance and intersection into a helper file
